% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAEapp.R
\name{runAEapp}
\alias{runAEapp}
\title{Run Adverse Events App}
\usage{
runAEapp(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Boolean logic to launch application in browser. Default to TRUE.}
}
\value{
An R Shiny Application.
}
\description{
Run Adverse Events App
}
